refactor: Money 클래스화
